paginables functions in queries.py
