+++ counterparty-core/counterpartycore/lib/api/routes.py
@@ -1,57 +1,54 @@
-from counterpartycore.lib import (
-    ledger,
-    transaction,
-)
-from counterpartycore.lib.api import util
+from counterpartycore.lib import transaction
+from counterpartycore.lib.api import queries, util
 from counterpartycore.lib.backend import addrindexrs, bitcoind
 
 # Define the API routes except root (`/`) defined in `api_server.py`
 ROUTES = util.prepare_routes(
     {
         ### /blocks ###
-        "/v2/blocks": ledger.get_blocks,
-        "/v2/blocks/<int:block_index>": ledger.get_block,
-        "/v2/blocks/<int:block_index>/transactions": ledger.get_transactions_by_block,
-        "/v2/blocks/<int:block_index>/events": ledger.get_events_by_block,
-        "/v2/blocks/<int:block_index>/events/counts": ledger.get_event_counts_by_block,
-        "/v2/blocks/<int:block_index>/events/<event>": ledger.get_events_by_block_and_event,
-        "/v2/blocks/<int:block_index>/credits": ledger.get_credits_by_block,
-        "/v2/blocks/<int:block_index>/debits": ledger.get_debits_by_block,
-        "/v2/blocks/<int:block_index>/expirations": ledger.get_expirations,
-        "/v2/blocks/<int:block_index>/cancels": ledger.get_cancels,
-        "/v2/blocks/<int:block_index>/destructions": ledger.get_destructions,
-        "/v2/blocks/<int:block_index>/issuances": ledger.get_issuances_by_block,
-        "/v2/blocks/<int:block_index>/sends": ledger.get_sends_by_block,
-        "/v2/blocks/<int:block_index>/dispenses": ledger.get_dispenses_by_block,
-        "/v2/blocks/<int:block_index>/sweeps": ledger.get_sweeps_by_block,
+        "/v2/blocks": queries.get_blocks,
+        "/v2/blocks/<int:block_index>": queries.get_block_by_height,
+        "/v2/blocks/<int:block_index>/transactions": queries.get_transactions_by_block,
+        "/v2/blocks/<int:block_index>/events": queries.get_events_by_block,
+        "/v2/blocks/<int:block_index>/events/counts": queries.get_event_counts_by_block,
+        "/v2/blocks/<int:block_index>/events/<event>": queries.get_events_by_block_and_event,
+        "/v2/blocks/<int:block_index>/credits": queries.get_credits_by_block,
+        "/v2/blocks/<int:block_index>/debits": queries.get_debits_by_block,
+        "/v2/blocks/<int:block_index>/expirations": queries.get_expirations,
+        "/v2/blocks/<int:block_index>/cancels": queries.get_cancels,
+        "/v2/blocks/<int:block_index>/destructions": queries.get_destructions,
+        "/v2/blocks/<int:block_index>/issuances": queries.get_issuances_by_block,
+        "/v2/blocks/<int:block_index>/sends": queries.get_sends_by_block,
+        "/v2/blocks/<int:block_index>/dispenses": queries.get_dispenses_by_block,
+        "/v2/blocks/<int:block_index>/sweeps": queries.get_sweeps_by_block,
         ### /transactions ###
         "/v2/transactions/info": transaction.info,
         "/v2/transactions/unpack": transaction.unpack,
         "/v2/transactions/<int:tx_index>": transaction.get_transaction_by_tx_index,
         "/v2/transactions/<tx_hash>": transaction.get_transaction_by_hash,
-        "/v2/transactions/<int:tx_index>/events": ledger.get_events_by_transaction_index,
-        "/v2/transactions/<tx_hash>/events": ledger.get_events_by_transaction_hash,
-        "/v2/transactions/<int:tx_index>/events/<event>": ledger.get_events_by_transaction_index_and_event,
-        "/v2/transactions/<tx_hash>/events/<event>": ledger.get_events_by_transaction_hash_and_event,
+        "/v2/transactions/<int:tx_index>/events": queries.get_events_by_transaction_index,
+        "/v2/transactions/<tx_hash>/events": queries.get_events_by_transaction_hash,
+        "/v2/transactions/<int:tx_index>/events/<event>": queries.get_events_by_transaction_index_and_event,
+        "/v2/transactions/<tx_hash>/events/<event>": queries.get_events_by_transaction_hash_and_event,
         ### /addresses ###
-        "/v2/addresses/<address>/balances": ledger.get_address_balances,
-        "/v2/addresses/<address>/balances/<asset>": ledger.get_balance_by_address_and_asset,
-        "/v2/addresses/<address>/credits": ledger.get_credits_by_address,
-        "/v2/addresses/<address>/debits": ledger.get_debits_by_address,
-        "/v2/addresses/<address>/bets": ledger.get_bet_by_feed,
-        "/v2/addresses/<address>/broadcasts": ledger.get_broadcasts_by_source,
-        "/v2/addresses/<address>/burns": ledger.get_burns_by_address,
-        "/v2/addresses/<address>/sends": ledger.get_send_by_address,
-        "/v2/addresses/<address>/receives": ledger.get_receive_by_address,
-        "/v2/addresses/<address>/sends/<asset>": ledger.get_send_by_address_and_asset,
-        "/v2/addresses/<address>/receives/<asset>": ledger.get_receive_by_address_and_asset,
-        "/v2/addresses/<address>/dispensers": ledger.get_dispensers_by_address,
-        "/v2/addresses/<address>/dispensers/<asset>": ledger.get_dispensers_by_address_and_asset,
-        "/v2/addresses/<address>/dispenses/sends": ledger.get_dispenses_by_source,
-        "/v2/addresses/<address>/dispenses/receives": ledger.get_dispenses_by_destination,
-        "/v2/addresses/<address>/dispenses/sends/<asset>": ledger.get_dispenses_by_source_and_asset,
-        "/v2/addresses/<address>/dispenses/receives/<asset>": ledger.get_dispenses_by_destination_and_asset,
-        "/v2/addresses/<address>/sweeps": ledger.get_sweeps_by_address,
+        "/v2/addresses/<address>/balances": queries.get_address_balances,
+        "/v2/addresses/<address>/balances/<asset>": queries.get_balance_by_address_and_asset,
+        "/v2/addresses/<address>/credits": queries.get_credits_by_address,
+        "/v2/addresses/<address>/debits": queries.get_debits_by_address,
+        "/v2/addresses/<address>/bets": queries.get_bet_by_feed,
+        "/v2/addresses/<address>/broadcasts": queries.get_broadcasts_by_source,
+        "/v2/addresses/<address>/burns": queries.get_burns_by_address,
+        "/v2/addresses/<address>/sends": queries.get_send_by_address,
+        "/v2/addresses/<address>/receives": queries.get_receive_by_address,
+        "/v2/addresses/<address>/sends/<asset>": queries.get_send_by_address_and_asset,
+        "/v2/addresses/<address>/receives/<asset>": queries.get_receive_by_address_and_asset,
+        "/v2/addresses/<address>/dispensers": queries.get_dispensers_by_address,
+        "/v2/addresses/<address>/dispensers/<asset>": queries.get_dispensers_by_address_and_asset,
+        "/v2/addresses/<address>/dispenses/sends": queries.get_dispenses_by_source,
+        "/v2/addresses/<address>/dispenses/receives": queries.get_dispenses_by_destination,
+        "/v2/addresses/<address>/dispenses/sends/<asset>": queries.get_dispenses_by_source_and_asset,
+        "/v2/addresses/<address>/dispenses/receives/<asset>": queries.get_dispenses_by_destination_and_asset,
+        "/v2/addresses/<address>/sweeps": queries.get_sweeps_by_address,
         ### /addresses/<address>/compose/ ###
         "/v2/addresses/<address>/compose/bet": transaction.compose_bet,
         "/v2/addresses/<address>/compose/broadcast": transaction.compose_broadcast,
@@ -67,39 +64,39 @@ ROUTES = util.prepare_routes(
         "/v2/addresses/<address>/compose/send": transaction.compose_send,
         "/v2/addresses/<address>/compose/sweep": transaction.compose_sweep,
         ### /assets ###
-        "/v2/assets": ledger.get_valid_assets,
-        "/v2/assets/<asset>": ledger.get_asset_info,
-        "/v2/assets/<asset>/balances": ledger.get_asset_balances,
-        "/v2/assets/<asset>/balances/<address>": ledger.get_balance_by_address_and_asset,
-        "/v2/assets/<asset>/orders": ledger.get_orders_by_asset,
-        "/v2/assets/<asset>/credits": ledger.get_credits_by_asset,
-        "/v2/assets/<asset>/debits": ledger.get_debits_by_asset,
-        "/v2/assets/<asset>/dividends": ledger.get_dividends,
-        "/v2/assets/<asset>/issuances": ledger.get_issuances_by_asset,
-        "/v2/assets/<asset>/sends": ledger.get_sends_by_asset,
-        "/v2/assets/<asset>/dispensers": ledger.get_dispensers_by_asset,
-        "/v2/assets/<asset>/dispensers/<address>": ledger.get_dispensers_by_address_and_asset,
-        "/v2/assets/<asset>/holders": ledger.get_asset_holders,
-        "/v2/assets/<asset>/dispenses": ledger.get_dispenses_by_asset,
+        "/v2/assets": queries.get_valid_assets,
+        "/v2/assets/<asset>": queries.get_asset_info,
+        "/v2/assets/<asset>/balances": queries.get_asset_balances,
+        "/v2/assets/<asset>/balances/<address>": queries.get_balance_by_address_and_asset,
+        "/v2/assets/<asset>/orders": queries.get_orders_by_asset,
+        "/v2/assets/<asset>/credits": queries.get_credits_by_asset,
+        "/v2/assets/<asset>/debits": queries.get_debits_by_asset,
+        "/v2/assets/<asset>/dividends": queries.get_dividends,
+        "/v2/assets/<asset>/issuances": queries.get_issuances_by_asset,
+        "/v2/assets/<asset>/sends": queries.get_sends_by_asset,
+        "/v2/assets/<asset>/dispensers": queries.get_dispensers_by_asset,
+        "/v2/assets/<asset>/dispensers/<address>": queries.get_dispensers_by_address_and_asset,
+        "/v2/assets/<asset>/holders": queries.get_asset_holders,
+        "/v2/assets/<asset>/dispenses": queries.get_dispenses_by_asset,
         ### /orders ###
-        "/v2/orders/<order_hash>": ledger.get_order,
-        "/v2/orders/<order_hash>/matches": ledger.get_order_matches_by_order,
-        "/v2/orders/<order_hash>/btcpays": ledger.get_btcpays_by_order,
-        "/v2/orders/<asset1>/<asset2>": ledger.get_orders_by_two_assets,
+        "/v2/orders/<order_hash>": queries.get_order,
+        "/v2/orders/<order_hash>/matches": queries.get_order_matches_by_order,
+        "/v2/orders/<order_hash>/btcpays": queries.get_btcpays_by_order,
+        "/v2/orders/<asset1>/<asset2>": queries.get_orders_by_two_assets,
         ### /bets ###
-        "/v2/bets/<bet_hash>": ledger.get_bet,
-        "/v2/bets/<bet_hash>/matches": ledger.get_bet_matches_by_bet,
-        "/v2/bets/<bet_hash>/resolutions": ledger.get_resolutions_by_bet,
+        "/v2/bets/<bet_hash>": queries.get_bet,
+        "/v2/bets/<bet_hash>/matches": queries.get_bet_matches_by_bet,
+        "/v2/bets/<bet_hash>/resolutions": queries.get_resolutions_by_bet,
         ### /burns ###
-        "/v2/burns": ledger.get_all_burns,
+        "/v2/burns": queries.get_all_burns,
         ### /dispensers ###
-        "/v2/dispensers/<dispenser_hash>": ledger.get_dispenser_info_by_hash,
-        "/v2/dispensers/<dispenser_hash>/dispenses": ledger.get_dispenses_by_dispenser,
+        "/v2/dispensers/<dispenser_hash>": queries.get_dispenser_info_by_hash,
+        "/v2/dispensers/<dispenser_hash>/dispenses": queries.get_dispenses_by_dispenser,
         ### /events ###
-        "/v2/events": ledger.get_all_events,
-        "/v2/events/<int:event_index>": ledger.get_event_by_index,
-        "/v2/events/counts": ledger.get_all_events_counts,
-        "/v2/events/<event>": ledger.get_events_by_name,
+        "/v2/events": queries.get_all_events,
+        "/v2/events/<int:event_index>": queries.get_event_by_index,
+        "/v2/events/counts": queries.get_all_events_counts,
+        "/v2/events/<event>": queries.get_events_by_name,
         ### /healthz ###
         "/v2/healthz": util.check_server_health,
         "/healthz": util.check_server_health,
@@ -112,9 +109,9 @@ ROUTES = util.prepare_routes(
         "/v2/bitcoin/estimatesmartfee": bitcoind.fee_per_kb,
         "/v2/bitcoin/transactions": bitcoind.sendrawtransaction,
         ### /mempool ###
-        "/v2/mempool/events": ledger.get_all_mempool_events,
-        "/v2/mempool/events/<event>": ledger.get_mempool_events_by_name,
-        "/v2/mempool/transactions/<tx_hash>/events": ledger.get_mempool_events_by_tx_hash,
+        "/v2/mempool/events": queries.get_all_mempool_events,
+        "/v2/mempool/events/<event>": queries.get_mempool_events_by_name,
+        "/v2/mempool/transactions/<tx_hash>/events": queries.get_mempool_events_by_tx_hash,
         ### API v1 ###
         "/": util.redirect_to_rpc_v1,
         "/v1/": util.redirect_to_rpc_v1,
